Move proofs example to example dir
